+++ R/tq_mutate.R
@@ -7,20 +7,15 @@
 #' @name tq_mutate
 #'
 #' @param data A `tibble` (tidy data frame) of data typically from [tq_get()].
-#' @param ohlc_fun The `quantmod` function that identifies columns to pass to
-#' the mutatation function. OHLCV is `quantmod` terminology for
-#' open, high, low, close, and volume. Options include c(Op, Hi, Lo, Cl, Vo, Ad,
-#' HLC, OHLC, OHLCV).
+#' @param select The columns to send to the mutation function.
 #' @param x,y Parameters used with `_xy` that consist of column names of variables
 #' to be passed to the mutatation function (instead of OHLC functions).
-#' @param select Optional parameter used with `_data` to select specific columns
-#' to send to the mutation function
 #' @param mutate_fun The mutation function from either the `xts`,
 #' `quantmod`, or `TTR` package. Execute `tq_mutate_fun_options()`
 #' to see the full list of options by package.
 #' @param col_rename A string or character vector containing names that can be used
 #' to quickly rename columns.
-#' @param transform_fun Deprecated. Use `mutate_fun`.
+#' @param ohlc_fun Deprecated. Use `select`.
 #' @param ... Additional parameters passed to the appropriate mutatation
 #' function.
 #'
@@ -36,34 +31,25 @@
 #' changing periodicity where the new columns would not have the same number of rows
 #' as the original tibble.
 #'
-#' `ohlc_fun` is one of the various `quantmod` Open, High, Low, Close (OHLC) functions.
-#' The function returns a column or set of columns from `data` that are passed to the
-#' `mutate_fun`. In Example 1 below, `Cl` returns the "close" price and sends
-#' this to the mutate function, `periodReturn`.
+#' `select` specifies the columns that get passed to the mutation function. Select works
+#' as a more flexible version of the OHLC extractor functions from `quantmod` where
+#' non-OHLC data works as well. When `select` is `NULL`, all columns are selected.
+#' In Example 1 below, `close` returns the "close" price and sends this to the
+#' mutate function, `periodReturn`.
 #'
 #' `mutate_fun` is the function that performs the work. In Example 1, this
 #' is `periodReturn`, which calculates the period returns. The `...`
-#' functions are additional arguments passed to the `mutate_fun`. Think of
-#' the whole operation in Example 1 as the close price, obtained by `ohlc_fun = Cl`,
-#' is being sent to the `periodReturn` function along
+#' are additional arguments passed to the `mutate_fun`. Think of
+#' the whole operation in Example 1 as the close price, obtained by `select = close`,
+#' being sent to the `periodReturn` function along
 #' with additional arguments defining how to perform the period return, which
 #' includes `period = "daily"` and `type = "log"`.
-#'
-#' `tq_mutate_xy` and `tq_transmute_xy` are designed to enable working with (1) mutatation
-#' functions that require two primary inputs (e.g. EVWMA, VWAP, etc) and (2) data
-#' that is not in OHLC format. Example 2 shows the first benefit in action:
-#' using the EVWMA function that uses volume to defind the moving average period.
-#' The two variables do not fall into a single OHLC code (i.e. CV does not exist).
-#' The xy form gets us out of this problem. Example 3 shows the second benefit
-#' in action: Some functions are useful to non-OHLC data, and defining x = price
-#' allows us to mutate WTI crude prices from daily to monthly periodicity.
-#'
-#' `tq_mutate_data` and `tq_transmute_data` are designed to enable working with
-#' functions that require "data" as the input (i.e. `rollapply`). This allows
-#' working with the `FUN` argument and apply `by.column = FALSE`, which is
-#' specifically designed for handling complex functions that require multiple
-#' column inputs into the `FUN` function (e.g. rolling regressions using `lm`).
 #' Example 4 shows how to apply a rolling regression.
+#'
+#' `tq_mutate_xy` and `tq_transmute_xy` are designed to enable working with mutatation
+#' functions that require two primary inputs (e.g. EVWMA, VWAP, etc).
+#' Example 2 shows this benefit in action: using the EVWMA function that uses
+#' volume to define the moving average period.
 #'
 #' `tq_mutate_`, `tq_mutate_xy_`, `tq_transmute_`, and `tq_transmute_xy_`
 #' are setup for Non-Standard
@@ -91,7 +77,7 @@
 #'
 #' # Example 1: Return logarithmic daily returns using periodReturn()
 #' fb_stock_prices %>%
-#'     tq_mutate(ohlc_fun = Cl, mutate_fun = periodReturn,
+#'     tq_mutate(select = close, mutate_fun = periodReturn,
 #'               period = "daily", type = "log")
 #'
 #' # Example 2: Use tq_mutate_xy to use functions with two columns required
@@ -99,25 +85,25 @@
 #'     tq_mutate_xy(x = close, y = volume, mutate_fun = EVWMA,
 #'                  col_rename = "EVWMA")
 #'
-#' # Example 3: Using tq_mutate_xy to work with non-OHLC data
+#' # Example 3: Using tq_mutate to work with non-OHLC data
 #' tq_get("DCOILWTICO", get = "economic.data") %>%
-#'     tq_mutate_xy(x = price, mutate_fun = lag.xts, k = 1, na.pad = TRUE)
+#'     tq_mutate(select = price, mutate_fun = lag.xts, k = 1, na.pad = TRUE)
 #'
-#' # Example 4: Using tq_mutate_data to apply a rolling regression
+#' # Example 4: Using tq_mutate to apply a rolling regression
 #' fb_returns <- fb_stock_prices %>%
-#'     tq_transmute(Ad, periodReturn, period = "monthly", col_rename = "fb.returns")
+#'     tq_transmute(adjusted, periodReturn, period = "monthly", col_rename = "fb.returns")
 #' xlk_returns <- tq_get("XLK", from = "2016-01-01", to = "2016-12-31") %>%
-#'     tq_transmute(Ad, periodReturn, period = "monthly", col_rename = "xlk.returns")
+#'     tq_transmute(adjusted, periodReturn, period = "monthly", col_rename = "xlk.returns")
 #' returns_combined <- left_join(fb_returns, xlk_returns, by = "date")
 #' regr_fun <- function(data) {
 #'     coef(lm(fb.returns ~ xlk.returns, data = as_data_frame(data)))
 #' }
 #' returns_combined %>%
-#'     tq_mutate_data(mutate_fun = rollapply,
-#'                    width      = 6,
-#'                    FUN        = regr_fun,
-#'                    by.column  = FALSE,
-#'                    col_rename = c("coef.0", "coef.1"))
+#'     tq_mutate(mutate_fun = rollapply,
+#'               width      = 6,
+#'               FUN        = regr_fun,
+#'               by.column  = FALSE,
+#'               col_rename = c("coef.0", "coef.1"))
 #'
 #' # Example 5: Non-standard evaluation:
 #' # Programming with tq_mutate_() and tq_mutate_xy_()
@@ -130,11 +116,32 @@ NULL
 
 #' @rdname tq_mutate
 #' @export
-tq_mutate <- function(data, ohlc_fun = OHLCV, mutate_fun, col_rename = NULL, ...) {
+tq_mutate <- function(data, select = NULL, mutate_fun, col_rename = NULL, ohlc_fun = NULL, ...) {
+
+    # Deprecate ohlc_fun in favor of select
+    if (!missing(ohlc_fun)) {
+        warning("Argument ohlc_fun is deprecated; please use select instead.",
+                call. = FALSE)
+
+        # As text
+        ohlc_string <- lazyeval::expr_text(ohlc_fun)
+
+        # Find select equivalent or die trying
+        select <- map_ohlc_to_select(ohlc_string)
+
+        # NSE and return
+        return(
+            tq_mutate_(data       = data,
+                       select     = select,
+                       mutate_fun = lazyeval::expr_text(mutate_fun),
+                       col_rename = col_rename,
+                       ...        = ...)
+        )
+    }
 
     # NSE
     tq_mutate_(data       = data,
-               ohlc_fun   = lazyeval::expr_text(ohlc_fun),
+               select     = lazyeval::expr_text(select),
                mutate_fun = lazyeval::expr_text(mutate_fun),
                col_rename = col_rename,
                ...        = ...)
@@ -142,25 +149,25 @@ tq_mutate <- function(data, ohlc_fun = OHLCV, mutate_fun, col_rename = NULL, ...
 
 #' @rdname tq_mutate
 #' @export
-tq_mutate_ <- function(data, ohlc_fun = "OHLCV", mutate_fun, col_rename = NULL, ...) {
+tq_mutate_ <- function(data, select = NULL, mutate_fun, col_rename = NULL, ...) {
     UseMethod("tq_mutate_", data)
 }
 
 # tq_mutate method dispatch --------------------------------------------------------------------------------
 
 #' @export
-tq_mutate_.default <- function(data, ohlc_fun = "OHLCV", mutate_fun, col_rename = NULL, ...) {
+tq_mutate_.default <- function(data, select = NULL, mutate_fun, col_rename = NULL, ...) {
 
     # Error message
     stop("data must be a tibble or data.frame object")
 }
 
 #' @export
-tq_mutate_.tbl_df <- function(data, ohlc_fun = "OHLCV", mutate_fun, col_rename = NULL, ...) {
+tq_mutate_.tbl_df <- function(data, select = NULL, mutate_fun, col_rename = NULL, ...) {
 
     # Get transformation
     ret <- tq_transmute_(data          = data,
-                         ohlc_fun      = ohlc_fun,
+                         select        = select,
                          mutate_fun    = mutate_fun,
                          col_rename    = col_rename,
                          ...           = ...)
@@ -169,14 +176,14 @@ tq_mutate_.tbl_df <- function(data, ohlc_fun = "OHLCV", mutate_fun, col_rename =
 }
 
 #' @export
-tq_mutate_.data.frame <- function(data, ohlc_fun = "OHLCV", mutate_fun, col_rename = NULL, ...) {
+tq_mutate_.data.frame <- function(data, select = NULL, mutate_fun, col_rename = NULL, ...) {
 
     # Convert data.frame to tibble
     data <- as_tibble(data)
 
     # Get transformation
     ret <- tq_transmute_(data          = data,
-                         ohlc_fun      = ohlc_fun,
+                         select        = select,
                          mutate_fun    = mutate_fun,
                          col_rename    = col_rename,
                          ...           = ...)
@@ -241,64 +248,6 @@ tq_mutate_xy_.data.frame <- function(data, x, y = NULL, mutate_fun, col_rename =
                             mutate_fun    = mutate_fun,
                             col_rename    = col_rename,
                             ...           = ...)
-
-    merge_two_tibbles(tib1 = data, tib2 = ret, mutate_fun)
-}
-
-# tq_mutate_data ------------------------------------------------------------------------------------------------
-
-#' @rdname tq_mutate
-#' @export
-tq_mutate_data <- function(data, select = NULL, mutate_fun, col_rename = NULL, ...) {
-
-    # NSE
-    tq_mutate_data_(data       = data,
-                    select     = lazyeval::expr_text(select),
-                    mutate_fun = lazyeval::expr_text(mutate_fun),
-                    col_rename = col_rename,
-                    ...        = ...)
-}
-
-#' @rdname tq_mutate
-#' @export
-tq_mutate_data_ <- function(data, select = NULL, mutate_fun, col_rename = NULL, ...) {
-    UseMethod("tq_mutate_data_", data)
-}
-
-# tq_mutate_data method dispatch --------------------------------------------------------------------------------
-
-#' @export
-tq_mutate_data_.default <- function(data, select = NULL, mutate_fun, col_rename = NULL, ...) {
-
-    # Error message
-    stop("data must be a tibble or data.frame object")
-}
-
-#' @export
-tq_mutate_data_.tbl_df <- function(data, select = NULL, mutate_fun, col_rename = NULL, ...) {
-
-    # Get transformation
-    ret <- tq_transmute_data_(data          = data,
-                              select        = select,
-                              mutate_fun    = mutate_fun,
-                              col_rename    = col_rename,
-                              ...           = ...)
-
-    merge_two_tibbles(tib1 = data, tib2 = ret, mutate_fun)
-}
-
-#' @export
-tq_mutate_data_.data.frame <- function(data, select = NULL, mutate_fun, col_rename = NULL, ...) {
-
-    # Convert data.frame to tibble
-    data <- as_tibble(data)
-
-    # Get transformation
-    ret <- tq_transmute_data_(data          = data,
-                              select        = select,
-                              mutate_fun    = mutate_fun,
-                              col_rename    = col_rename,
-                              ...           = ...)
 
     merge_two_tibbles(tib1 = data, tib2 = ret, mutate_fun)
 }